refactor: RacingCar module separated, singleton pattern
