+++ package.json
@@ -44,5 +44,6 @@
     "presets": [
       "@babel/preset-env"
     ]
-  }
+  },
+  "type": "module"
 }
